Convert gulpfile to ES6 and make it modular
This will allow the gulpfile to scale better, and facilitate
migration of the old build tasks from Grunt.
