+++ gulp-tasks/build-assets.js
@@ -1,0 +1,27 @@
+import es from 'event-stream';
+
+import gulp from 'gulp';
+import concat from 'gulp-concat';
+import replace from 'gulp-replace';
+import htmlreplace from 'gulp-html-replace';
+
+// Concatenates CSS files, rewrites relative paths to Bootstrap fonts, copies Bootstrap fonts
+gulp.task('css', () => {
+    var bowerCss = gulp.src('src/bower_modules/components-bootstrap/css/bootstrap.min.css')
+            .pipe(replace(/url\((')?\.\.\/fonts\//g, 'url($1fonts/')),
+        appCss = gulp.src('src/css/*.css'),
+        combinedCss = es.concat(bowerCss, appCss).pipe(concat('css.css')),
+        fontFiles = gulp.src('./src/bower_modules/components-bootstrap/fonts/*', { base: './src/bower_modules/components-bootstrap/' });
+    return es.concat(combinedCss, fontFiles)
+        .pipe(gulp.dest('./dist/'));
+});
+
+// Copies index.html, replacing <script> and <link> tags to reference production URLs
+gulp.task('html', () => {
+    return gulp.src('./src/index.html')
+        .pipe(htmlreplace({
+            'css': 'css.css',
+            'js': 'scripts.js'
+        }))
+        .pipe(gulp.dest('./dist/'));
+});
